organizacao final
