+++ semana9/astromatch/src/kiddos/Home/Buttons.js
@@ -5,10 +5,12 @@ import coloredYes from '../../assets/painted-like.svg'
 import coloredNo from '../../assets/painted-dislike.svg'
 import styled from 'styled-components'
 
-
 const Icon = styled.img`
   width: 20%;
   cursor: pointer;
+  &:hover {
+    width: 21%
+  }
 `
 
 const ButtonContainer = styled.div`
